requirements.txt
PyYAML==6.0.2
XlsxWriter==3.2.3
allpairspy==2.5.1
annotated-types==0.7.0
anyio==4.9.0
approval_utilities==14.5.0
approvaltests==14.5.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
cryptography==44.0.3
distro==1.9.0
empty-files==0.0.9
fastapi==0.115.12
feedparser==6.0.11
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jiter==0.9.0
jsonschema-specifications==2025.4.1
jsonschema==4.23.0
lxml==5.4.0
mock==5.2.0
mrjob==0.7.4
mypy_extensions==1.1.0
numpy==2.2.5
openai==1.78.1
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.5.0
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pypdfium2==4.30.1
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
python-pptx==1.0.2
pytrends==4.9.2
pytz==2025.2
referencing==0.36.2
reportlab==4.4.0
requests==2.32.3
rpds-py==0.24.0
setuptools==80.4.0
sgmllib3k==1.0.0
six==1.17.0
slack-bolt==1.18.1
slack_sdk==3.35.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
wheel==0.45.1
pydantic-settings==2.5.2
aiohttp>=3.8.6,<4.0
fpdf>=1.7.2